moved input to input controller module #495
